Backend overhaul

+++ src/envoy/constants.rs
@@ -1,6 +1,6 @@
 //! Constants used across the envoy system
-pub const NUMBER_OF_MODULES: i32 = 12;
-pub const MUTANT_ID: i32 = 11;
+pub const NUMBER_OF_MODULES: usize = 12;
+pub const MUTANT_ID: usize = 11;
 pub const PROTOCOL: &str = "TCP";
 pub const DATA_ROUTER_LISTENER_PORT: i32 = 46005;
 pub const DATA_EXPORTER_LISTENER_PORT: i32 = 46007;
